Fix cloudfront key pathing for development

+++ config/application.rb
@@ -24,9 +24,11 @@ module Bps
       Devise::Mailer.layout 'mailer'
     end
 
-    k = File.open(Rails.application.secrets[:cf_private_key_path], 'w+')
-    File.chmod(0600, k)
-    k.write(ENV['CLOUDFRONT_PRIVATE_KEY'])
-    k.close
+    unless File.exist?(Rails.application.secrets[:cf_private_key_path])
+      k = File.open(Rails.application.secrets[:cf_private_key_path], 'w+')
+      File.chmod(0600, k)
+      k.write(ENV['CLOUDFRONT_PRIVATE_KEY'])
+      k.close
+    end
   end
 end
